+++ core/analysis/src/rules/mod.rs
@@ -28,6 +28,7 @@ pub mod struct_uzumaki_as_argument;
 pub mod top_level_const;
 pub mod uninitialized_variable;
 pub mod uzumaki_in_reassignment;
+pub mod uzumaki_on_compound_array_element;
 pub mod uzumaki_on_compound_field;
 pub mod uzumaki_on_nested_struct;
 pub mod uzumaki_on_struct_in_array;
@@ -65,6 +66,7 @@ use struct_uzumaki_as_argument::StructUzumakiAsArgument;
 use top_level_const::TopLevelConstNotSupported;
 use uninitialized_variable::UninitializedVariable;
 use uzumaki_in_reassignment::UzumakiInReassignment;
+use uzumaki_on_compound_array_element::UzumakiOnCompoundArrayElement;
 use uzumaki_on_compound_field::UzumakiOnCompoundField;
 use uzumaki_on_nested_struct::UzumakiOnNestedStruct;
 use uzumaki_on_struct_in_array::UzumakiOnStructInArray;
@@ -117,5 +119,6 @@ pub fn all_rules() -> &'static [&'static dyn crate::rule::Rule] {
         &ArrayIndexConstOob,
         &UzumakiOnCompoundField,
         &StructUzumakiAsArgument,
+        &UzumakiOnCompoundArrayElement,
     ]
 }